updates nav buttons to use same css. refactors them to one directory
