--- src/components/internalNavButton/internalNavButton.scss
@@ -1,15 +1,0 @@
-.navigation-button {
-    font-size: 0.875em;
-    padding: 10px;
-    color: #7000ff;
-    letter-spacing: 2px;
-    text-transform: uppercase;
-    border-radius: 3px;
-    border: 2px solid #7000ff;
-    margin: 3rem auto;
-}
-
-.navigation-button:hover {
-    color: white;
-    background: #7000ff;
-}
